refactor: add isOffline util and rename downcaseKeys files
