+++ libs/utils/src/index.ts
@@ -1,2 +1,3 @@
-export * from './downcaseKeys';
+export * from './downcase-keys';
 export * from './invert';
+export * from './is-offline';
